Add unit tests for Subscriptions service
